+++ droidboy/src/main/java/com/appboy/sample/DroidboyNotificationFactory.java
@@ -2,7 +2,6 @@ package com.appboy.sample;
 
 import android.app.Notification;
 import android.content.Context;
-import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
 import android.support.v4.app.NotificationCompat;
@@ -17,9 +16,7 @@ public class DroidboyNotificationFactory implements IAppboyNotificationFactory {
   public Notification createNotification(AppboyConfigurationProvider appConfigurationProvider,
                                          Context context, Bundle notificationExtras, Bundle appboyExtras) {
     NotificationCompat.Builder notificationBuilder = AppboyNotificationFactory.getInstance().populateNotificationBuilder(appConfigurationProvider, context, notificationExtras, appboyExtras);
-    if (Build.VERSION.SDK_INT > Build.VERSION_CODES.HONEYCOMB_MR1) {
-      notificationBuilder.setSound(Settings.System.DEFAULT_NOTIFICATION_URI);
-    }
+    notificationBuilder.setSound(Settings.System.DEFAULT_NOTIFICATION_URI);
     return notificationBuilder.build();
   }
 }